+++ tracer/src/Datadog.Trace/Configuration/ExporterSettings.cs
@@ -8,8 +8,10 @@
 using System;
 using System.Collections.Generic;
 using Datadog.Trace.Agent;
+using Datadog.Trace.AlwaysOnProfiler;
 using Datadog.Trace.Configuration.Helpers;
 using Datadog.Trace.SignalFx.Metrics;
+using Datadog.Trace.Vendors.Serilog;
 using MetricsTransportType = Datadog.Trace.Vendors.StatsdClient.Transport.TransportType;
 
 namespace Datadog.Trace.Configuration
@@ -96,6 +98,21 @@ namespace Datadog.Trace.Configuration
                 defaultTo: 500,
                 validators: (val) => val > 0);
             SyncExport = source?.GetBool(ConfigurationKeys.TraceSyncExport) ?? false;
+
+            var stringProfileExportFormat = source?.GetString(ConfigurationKeys.AlwaysOnProfiler.ExportFormat);
+            if (string.IsNullOrEmpty(stringProfileExportFormat))
+            {
+                ProfilerExportFormat = ProfilerExportFormat.Pprof;
+            }
+            else if (Enum.TryParse(stringProfileExportFormat, out ProfilerExportFormat profilerExportFormat))
+            {
+                ProfilerExportFormat = profilerExportFormat;
+            }
+            else
+            {
+                Log.Logger.Warning("Unknown {0} passed: {1}, using {2} as a default", ConfigurationKeys.AlwaysOnProfiler.ExportFormat, stringProfileExportFormat, ProfilerExportFormat.Pprof);
+                ProfilerExportFormat = ProfilerExportFormat.Pprof;
+            }
         }
 
         /// <summary>
@@ -175,6 +192,11 @@ namespace Datadog.Trace.Configuration
         public MetricsExporterType MetricsExporter { get; set; }
 
         internal List<string> ValidationWarnings { get; }
+
+        /// <summary>
+        /// Gets or sets the format in which the profiling data will be exported.
+        /// </summary>
+        internal ProfilerExportFormat ProfilerExportFormat { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum number of closed spans in a trace before it's partially flushed