+++ tracer/src/Datadog.Trace/Configuration/TracerSettings.cs
@@ -204,17 +204,6 @@ namespace Datadog.Trace.Configuration
             // Filter out tags with empty keys or empty values, and trim whitespaces
             GrpcTags = InitializeHeaderTags(grpcTags, headerTagsNormalizationFixEnabled: true);
 
-            var stringProfileExportFormat = source?.GetString(ConfigurationKeys.AlwaysOnProfiler.ExportFormat);
-            if (Enum.TryParse(stringProfileExportFormat, out ProfilerExportFormat profilerExportFormat))
-            {
-                ProfilerExportFormat = profilerExportFormat;
-            }
-            else
-            {
-                Log.Logger.Warning("Unknown {0} passed: {1}, using {2} as a default", ConfigurationKeys.AlwaysOnProfiler.ExportFormat, stringProfileExportFormat, ProfilerExportFormat.Pprof);
-                ProfilerExportFormat = ProfilerExportFormat.Pprof;
-            }
-
             IsActivityListenerEnabled = source?.GetBool(ConfigurationKeys.FeatureFlags.ActivityListenerEnabled) ??
                     // default value
                     false;
@@ -526,8 +515,6 @@ namespace Datadog.Trace.Configuration
         /// Gets or sets a value indicating the trace methods configuration.
         /// </summary>
         internal string TraceMethods { get; set; }
-
-        internal ProfilerExportFormat ProfilerExportFormat { get; set; }
 
         /// <summary>
         /// Gets a value indicating whether the activity listener is enabled or not.